+++ internal/core/nginx/cfgfiles/file_provider.go
@@ -23,6 +23,7 @@ type SupportedFeatures struct {
 	StreamType  SupportType
 	RunCodeType SupportType
 	StatsType   SupportType
+	GRPCType    SupportType
 }
 
 type providerContext struct {